refactor(io): split io.rs into source/icc/exif submodules (#565) (#609)
* refactor(io): split io.rs into source/icc/exif submodules (#565)

Splits the 1772-line src/engine/io.rs into a directory module with one
file per concern, keeping every file under the 800-line ceiling from
CLAUDE.md:

- source.rs (244): Source enum + MmapGuard/MemoryGuard + load_file_safe
- icc.rs (589):    ICC extraction across JPEG/PNG/WebP/AVIF containers
- icc_tests.rs:    icc test body referenced via #[path] from icc.rs
- exif.rs (433):   EXIF metadata extraction + GPS IFD detection
- test_helpers.rs: shared image fixtures (cfg(test), pub(super))
- mod.rs (49):     public + crate-internal re-exports preserving the
                   historical crate::engine::io::* paths

All historical paths remain stable for engine.rs, tasks/*, firewall.rs,
and api.rs. No behavior change. Pure refactor: 443 Rust tests + full JS
suite pass.

* refactor(io): clean up duplicate icc tests and unsafe comments (#565)

Addresses rust-reviewer feedback on PR #609:

- Remove the duplicate `mod icc_tests` from src/engine.rs (lines 185-583
  in the previous form). The 25 tests were copied verbatim to
  src/engine/io/icc_tests.rs during the split but the original was not
  deleted, so every ICC test ran twice under different module paths.
  engine.rs drops from 720 → 222 lines.
- Drop the now-unused `extract_icc_ok` / `create_minimal_jpeg` /
  `create_minimal_webp` helpers from the engine.rs test module; the
  remaining `decode_tests` only need `create_test_image`, `create_png`,
  and `create_minimal_png`.
- Narrow `io/mod.rs` crate-internal re-exports. The `pub(crate) use`
  block for `IccExtractor` / `JpegIccExtractor` / `extract_icc_from_jpeg`
  / `validate_icc_profile` / `is_avif_data` / etc. had no in-crate
  callers after the engine.rs cleanup; remove it. The items remain
  reachable via `crate::engine::io::icc::*` when truly needed.
- Capitalise `// Safety:` → `// SAFETY:` in source.rs:177 so the comment
  matches the `undocumented_unsafe_blocks` lint convention.
- Add an explicit `// SAFETY:` comment on the inner `unsafe` block
  inside `AvifDecoderGuard::drop` in icc.rs that the outer guard's
  comment did not cover.

No behavior change. cargo fmt clean, clippy -D warnings clean, 418 Rust
tests pass, npm test passes.

* fix(io): remove unused test imports
